backend/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# AI/ML - Semantic Grading
sentence-transformers>=2.2.2
torch>=2.2.0
numpy>=1.26.0

# OCR - Handwriting Recognition
opencv-python-headless>=4.9.0
pytesseract>=0.3.10
Pillow>=10.2.0

# Analytics
pandas>=2.1.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
